KEYCLOAK-2045 Refactor Validator API
- Merged BuiltinValidators and ValidatorLookup into Validators facade class
