--- server-spi-private/src/main/java/org/keycloak/validate/ValidatorLookup.java
@@ -1,79 +1,0 @@
-/*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.keycloak.validate;
-
-import org.keycloak.models.KeycloakSession;
-import org.keycloak.models.KeycloakSessionFactory;
-import org.keycloak.validate.builtin.BuiltinValidators;
-
-/**
- * Allows to search for {@link Validator} implementations by id.
- */
-public class ValidatorLookup {
-
-    // TODO Validator lookup should be provided via KeycloakSession API later
-
-    /**
-     * Look-up up for a built-in or registered validator with the given validatorId.
-     *
-     * @param session the {@link KeycloakSession}
-     * @param id      the id of the validator
-     * @return the {@link Validator} or {@literal null}
-     */
-    public static Validator validator(KeycloakSession session, String id) {
-
-        // Fast-path for internal Validators
-        Validator validator = BuiltinValidators.getValidatorById(id);
-        if (validator != null) {
-            return validator;
-        }
-
-        if (session == null) {
-            return null;
-        }
-
-        // Lookup validator in registry
-        return session.getProvider(Validator.class, id);
-    }
-
-    /**
-     * Look-up for a built-in or registered validator with the given validatorId.
-     * <p>
-     * This is intended for users who want to dynamically create new {@link Validator} instances, validate
-     * {@link ValidatorConfig} configurations or create default configurations for a {@link Validator}.
-     *
-     * @param session the {@link KeycloakSession}
-     * @param id      the id of the validator
-     * @return the {@link Validator} or {@literal null}
-     */
-    public static ValidatorFactory validatorFactory(KeycloakSession session, String id) {
-
-        // Fast-path for internal Validators
-        ValidatorFactory factory = BuiltinValidators.getValidatorFactoryById(id);
-        if (factory != null) {
-            return factory;
-        }
-
-        if (session == null) {
-            return null;
-        }
-
-        // Lookup factory in registry
-        KeycloakSessionFactory sessionFactory = session.getKeycloakSessionFactory();
-        return (ValidatorFactory) sessionFactory.getProviderFactory(Validator.class, id);
-    }
-}
